Shortern the story of The Great Flood.
What
=
Shortern the story of The Great Flood, removing the beginning that
talked about the sons of God claiming the daughters of man as wives.

Why
=
I think it was included in the scope of this story to provide context to
what evil things were being done, but due to the varied opinions of
theologians on what it actually means I think it gets in the way of the
story snippet we're wanting to show here. A reader can always look at
the surronding text by using the link at the bottom of the page.