Separated `Passage`s from `Story`s and `Account`s.
What
=
Moved logic out of `Account` and put it in a new class called `Passage`.
The logic moved out was any logic about or that can be derived from the
passage or the source of the passage.

Why
=
To separate the passage as an independent piece of information that has
no dependencies on the web app at all, from the `Account` which is a
concept specific to this app. Will make it easier to reuse the passage
loading code in other projects and it may be broken out separately.