Supporting multiple versions of a passage.
What
=
Supporting multiple versions of a passage to test the CEV alongside the
ESV which is used today. This is for the English website only.

Why
=
To see if engagement increases when the translation used for stories is
easier to understand, and to confirm that switching would not negatively
impact engagement.