Refactored the Bible APIs.
What
===
Refactored the Bible APIs so that the Apis master passes in what
versions it wants from each API class. Each individual APIs class no
longer knows anything about locales.

Why
=
All the APIs classes didn't need to know about locales, and it was messy
that each had to know about it. This makes it easier to define new
versions for multiple languages, and for example to use the BiblesorgApi
to retrieve both CEVD and GNT for English.