Replaced middleman-cdn with simple CURL calls.
What
===
Replaced CDN invalidation using middleman-cdn with simple CURL calls to
the CloudFlare API.

Why
===
The project often changes so many files that individual file purge
limits are exceeded and therefore I need to invalidate the entire zone
instead. Invalidating an entire zone is a simple CURL, so there's no
need to use a gem that's more complicated and pulls in many dependencies.